+++ spider.go
@@ -54,6 +54,7 @@ v0.8.1;
 	updated default -delay to 10ms
 v0.9.0:
     added flag "-match" to only crawl URLs containing a specified keyword; https://github.com/cyclone-github/spider/issues/6
+	added notice to user if no URLs are crawled when using "-crawl 1 -match"
     exit early if zero URLs were crawled (no processing or file output)
     use custom User-Agent "Spider/0.9.0 (+https://github.com/cyclone-github/spider)"
     removed clearScreen function and its imports
@@ -365,7 +366,7 @@ func main() {
 		fmt.Fprintf(os.Stderr, "Base domain:\t%s\n", baseDomain)
 		fmt.Fprintf(os.Stderr, "Crawl depth:\t%d\n", *crawlFlag)
 		fmt.Fprintf(os.Stderr, "ngram len:\t%s\n", *ngramFlag)
-		fmt.Fprintf(os.Stderr, "Crawl delay:\t%dms (increase this to avoid rate limiting)\n", *delayFlag)
+		fmt.Fprintf(os.Stderr, "Crawl delay:\t%dms (increase to avoid rate limiting)\n", *delayFlag)
 		fmt.Fprintf(os.Stderr, "Timeout:\t%d sec\n", *timeoutFlag)
 	}
 
@@ -438,7 +439,11 @@ func main() {
 
 	// if nothing matched, exit early
 	if len(texts) == 0 {
+		time.Sleep(100)
 		fmt.Fprintln(os.Stderr, "No URLs crawled, exiting...") // boo, something went wrong!
+		if *crawlFlag == 1 {
+			fmt.Fprintln(os.Stderr, "Try increasing -crawl depth")
+		}
 		return
 	}